Added implementation to also store IPA pronunciations
I currently have an issue with testing this due to that fact that the JSON containing the information takes about half a second to load. This causes the list to be undefined when trying to call it. That means that I'm going to have to find a way to stall or keep the data structures in memory to be able to avoid this behavior  and test.