pushing to work on other computer
Got arpabetDict properly loaded by creating a script that just loads that variables, obfuscating the need for file loading. Got the background script passing messaged to the my content script. My next step is to use that information to update the pop-up and then fix the asynchronous calls to Merriam-Webster